# api/swagger/swagger.yaml
basePath: /api/v1
consumes: [application/json]
definitions:
  ErrorResponse:
    properties:
      message: {type: string}
    required: [message]
...
  appResponse:
    properties:
      code: {type: number}
      message: {type: string}
    required: [code]
...
  withingsCallback:
    properties:
      oauth_verifier: {type: string}
      uid: {type: string}
      userid: {type: string}
    type: object
host: localhost:5063
info: {title: IOTied App, version: 0.0.1}
paths:
...
  /withings/callback:
    post:
      description: Withings Auth Callback
      operationId: oauth_callback
      parameters:
      - description: Withings callback url
        in: body
        name: body
        required: true
        schema: {$ref: '#/definitions/withingsCallback'}
      responses:
        '200':
          description: Success
          schema: {$ref: '#/definitions/appResponse'}
        default:
          description: Error
          schema: {$ref: '#/definitions/ErrorResponse'}
      summary: Withings Auth Callback
      tags: [withings]
    x-swagger-router-controller: withings
  /withings/connect:
    get:
      description: Withings Auth Request
      operationId: authRequest
      responses:
        '200':
          description: Success
          schema: {$ref: '#/definitions/appResponse'}
        default:
          description: Error
          schema: {$ref: '#/definitions/ErrorResponse'}
      summary: Withings Auth Request
      tags: [withings]
    x-swagger-router-controller: withings
produces: [application/json]
schemes: [http, https]
swagger: '2.0'